:memo: Update Readmore examples
