+++ aksel.nav.no/website/pages/eksempler/search/error.tsx
@@ -23,4 +23,5 @@ export const Demo = {
 
 export const args = {
   index: 6,
+  desc: "Children kan erstatte knapp hvis man f.eks. trenger eget ikon eller 'type'-prop.",
 };
